+++ src/main/java/org/openlmis/requisition/web/FacilitySupportsProgramHelper.java
@@ -76,6 +76,15 @@ class FacilitySupportsProgramHelper {
         .orElse(null);
   }
 
+  /**
+   * Find a supported program entry from the given facilityId for the given program. If entry does
+   * not exist the {@code null} value will be returned.
+   */
+  SupportedProgramDto getSupportedProgram(UUID facilityId, UUID programId) {
+    FacilityDto facility = facilityReferenceDataService.findOne(facilityId);
+    return getSupportedProgram(facility, programId);
+  }
+
   private boolean isProgramSupported(List<SupportedProgramDto> supportedPrograms,
                                      UUID programId) {
     return supportedPrograms